table-browser: remove utils module
