+++ src/scripts/modules/table-browser/routes.js
@@ -1,5 +1,4 @@
 import Index from './react/Index';
-import {initLocalState} from './utils';
 import tableBrowserActions from './flux/actions';
 import storageActions from '../components/StorageActionCreators';
 import createActionsProvisioning from './actionsProvisioning';
@@ -8,9 +7,10 @@ export const PATH_PREFIX = 'tables';
 
 const requireDataFn = (routerState) => {
   const tableId = routerState.tableId;
+  const actions = createActionsProvisioning(tableId);
   return storageActions.loadTables().then( () => {
-    tableBrowserActions.setCurrentTableId(tableId, initLocalState(tableId));
-    createActionsProvisioning(tableId).loadAll();
+    tableBrowserActions.setCurrentTableId(tableId, actions.initLocalState(tableId));
+    actions.loadAll();
   }
   );
 };
